Automatisering/requirements.txt
Pillow
pytube
youtube_dl
PyPDF2
